fix : size 필드 오류 수정, 최신순 정렬 + id순 정렬 수정

+++ src/main/java/com/weady/weady/domain/board/dto/response/PageInfoDto.java
@@ -5,6 +5,6 @@ import lombok.Builder;
 @Builder
 public record PageInfoDto(
         Long nextCursor,
-        Long size,
+        Integer size,
         boolean hasNext
 ){}
